Fixed Cohen-Sutherland algo to handle straight lines & more (read desc for logs)
* Fixed Cohen-Sutherland algo when handling straight lines that clipped
* Has applied said algorithm in line drawing. Polygon will also be affected, since it is also using Line to draw each sides of Polygon
* Added getter getColor() for Line
* Changed yres & xres of DrawSurface to short, since it's value won't get past 2^16

+++ DrawSurface.hpp
@@ -6,12 +6,12 @@
 class DrawSurface
 {
   public:
-    int getXRes()
+    short getXRes()
     {
         return xres;
     }
 
-    int getYRes()
+    short getYRes()
     {
         return yres;
     }
@@ -20,8 +20,8 @@ class DrawSurface
     void virtual updateScreen() = 0;
 
   protected:
-    int yres;
-    int xres;
+    short yres;
+    short xres;
 };
 
 #endif
